bff/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
httpx==0.28.1
python-dotenv==1.2.2
